move LayerAdditions to disk and LayerIds to api
