+++ Cargo.toml
@@ -40,9 +40,9 @@ debug = 0
 
 [workspace.dependencies]
 #[public-storage]
-pometry-storage = { version = ">=0.8.1", path = "pometry-storage" }
+# pometry-storage = { version = ">=0.8.1", path = "pometry-storage" }
 #[private-storage]
-# pometry-storage = { path = "pometry-storage-private", package = "pometry-storage-private" }
+pometry-storage = { path = "pometry-storage-private", package = "pometry-storage-private", features = ["progress"]}
 async-graphql = { version = ">=7.0.5, <7.0.8", features = [
     "dynamic-schema",
 ] } # 7.0.8+ is borked, see https://github.com/async-graphql/async-graphql/issues/1586
