+++ raphtory-cypher/Cargo.toml
@@ -20,6 +20,7 @@ arrow.workspace = true
 arrow-buffer.workspace = true
 arrow-schema.workspace = true
 arrow-array.workspace = true
+tracing-subscriber.workspace = true
 
 pest.workspace = true
 pest_derive.workspace = true
