+++ raphtory-cypher/examples/raphtory_cypher.rs
@@ -14,7 +14,7 @@ mod cypher {
     use futures::{stream, StreamExt};
     use raphtory::{
         disk_graph::{graph_impl::ParquetLayerCols, DiskGraphStorage},
-        logging::global_info_logger,
+        logging::{global_debug_logger, global_info_logger, sysout_debug},
     };
     use raphtory_cypher::{run_cypher, run_cypher_to_streams, run_sql};
     use serde::{de::DeserializeOwned, Deserialize};
@@ -126,11 +126,11 @@ mod cypher {
 
     // #[tokio::main]
     pub async fn main() {
-        global_info_logger();
         let args = Args::parse();
 
         match args {
             Args::Query(args) => {
+                global_info_logger();
                 let graph =
                     DiskGraphStorage::load_from_dir(&args.graph_dir).expect("Failed to load graph");
 
@@ -145,7 +145,6 @@ mod cypher {
 
                     let now = std::time::Instant::now();
                     let batches = df.collect().await.unwrap();
-                    global_info_logger();
                     info!("Query execution time: {:?}", now.elapsed());
                     print_batches(&batches).expect("Failed to print batches");
                 } else {
@@ -161,6 +160,7 @@ mod cypher {
             }
 
             Args::Load(args) => {
+                sysout_debug();
                 let layers = args.layers;
                 let layer_parquet_cols = (0..layers.len())
                     .map(|layer_id| {
