+++ raphtory/src/algorithms/community_detection/louvain.rs
@@ -70,6 +70,7 @@ mod test {
         test_storage,
     };
     use proptest::prelude::*;
+    #[cfg(feature = "io")]
     use tracing::info;
 
     #[cfg(feature = "io")]
